backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
anthropic==0.34.2
python-dotenv==1.0.0
gunicorn==21.2.0
